get_new_batch_encode
